Refactoring to have a layered architectures (controllers/services/data layer)
